requirements.txt
absl-py==0.12.0
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
detectron2==0.1.3
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210508
google==3.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.8
iopath==0.1.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mistune==0.8.4
mock==4.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.52
packaging==20.9
pandocfilters==1.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.3.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.16.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.1
pyzmq==22.0.3
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.2.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==4.5.3
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
yacs==0.1.8
absl-py==0.12.0
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
detectron2
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210508
google==3.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.8
iopath==0.1.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mistune==0.8.4
mock==4.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.52
packaging==20.9
pandocfilters==1.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.3.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.16.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.1
pyzmq==22.0.3
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.2.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==4.5.3
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
yacs==0.1.8
